Refactor DetailsBox into shared component
