+++ components/measurement/DetailsBox.js
@@ -7,34 +7,52 @@ const DetailBoxLabel = styled(Text)`
   font-weight: 600;
 `
 
-const DetailsBox = ({
+export const DetailsBoxTable = ({
   title,
   items,
   bg
 }) => (
-  <Box width={1} bg={bg}>
-    <Box p={3}>
-      <Heading h={4}>{title}</Heading>
-      {items.map((item, index) =>
-        <Flex mb={1} key={index}>
-          <Box width={1/4}>
-            <DetailBoxLabel>{item.label}</DetailBoxLabel>
-          </Box>
-          <Box>
-            <Text>{item.value}</Text>
-          </Box>
-        </Flex>
-      )}
-    </Box>
-  </Box>
+  <DetailsBox title={title} bg={bg} content={
+    items.map((item, index) =>
+      <Flex flexWrap='wrap' key={index}>
+        <Box width={1/4}>
+          <DetailBoxLabel>{item.label}</DetailBoxLabel>
+        </Box>
+        <Box>
+          <Text>{item.value}</Text>
+        </Box>
+      </Flex>
+    )}
+  />
 )
 
-DetailsBox.propTypes = {
+DetailsBoxTable.propTypes = {
   title: PropTypes.string.isRequired,
   items: PropTypes.arrayOf(PropTypes.shape({
     label: PropTypes.string.isRequired,
     value: PropTypes.string
-  }))
+  })),
+  bg: PropTypes.string
 }
 
-export default DetailsBox
+const StyledDetailsBox = styled(Box)`
+  border: 2px solid ${props => props.theme.colors.gray2};
+`
+
+export const DetailsBox = ({ title, content, ...props }) => (
+  <StyledDetailsBox width={1} {...props} mb={3}>
+    {title &&
+      <Flex px={3} bg='gray2'>
+        <Heading h={4}>{title}</Heading>
+      </Flex>
+    }
+    <Box p={3} flexWrap='wrap'>
+      {content}
+    </Box>
+  </StyledDetailsBox>
+)
+
+DetailsBox.propTypes = {
+  title: PropTypes.string.isRequired,
+  content: PropTypes.element
+}
